# spec/swagger.yaml
definitions:
...
  bots.Bot:
    properties:
      avatar_url:
        type: string
      check_issue_count:
        type: integer
      check_state:
        type: string
      created_at:
        type: string
      display_name:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      metadata:
        additionalProperties: {}
        type: object
      owner_user_id:
        type: string
      status:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
...
  bots.UpdateBotRequest:
    properties:
      avatar_url:
        type: string
      display_name:
        type: string
      is_active:
        type: boolean
      metadata:
        additionalProperties: {}
        type: object
    type: object
...
  handlers.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
info:
  contact: {}
  title: Memoh API
  version: 1.0.0
paths:
...
  /bots/{id}:
    delete:
      description: Delete a bot user (owner/admin only)
      parameters:
      - description: Bot ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Delete bot
      tags:
      - bots
    get:
      description: Get a bot by ID (owner/admin only)
      parameters:
      - description: Bot ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/bots.Bot'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get bot details
      tags:
      - bots
    put:
      description: Update bot profile (owner/admin only)
      parameters:
      - description: Bot ID
        in: path
        name: id
        required: true
        type: string
      - description: Bot update payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/bots.UpdateBotRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/bots.Bot'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Update bot details
      tags:
      - bots
...
swagger: "2.0"